+++ yazi-cli/src/main.rs
@@ -78,10 +78,10 @@ async fn run() -> anyhow::Result<()> {
 			let mut pkg = package::Package::load().await?;
 			match cmd {
 				CommandPkg::Add { ids } => pkg.add_many(&ids).await?,
-				CommandPkg::Delete { ids } => pkg.delete_many(&ids).await?,
-				CommandPkg::Install => pkg.install().await?,
+				CommandPkg::Delete { ids, discard } => pkg.delete_many(&ids, discard).await?,
+				CommandPkg::Install { discard } => pkg.install(discard).await?,
 				CommandPkg::List => pkg.print()?,
-				CommandPkg::Upgrade { ids } => pkg.upgrade_many(&ids).await?,
+				CommandPkg::Upgrade { ids, discard } => pkg.upgrade_many(&ids, discard).await?,
 			}
 		}